commune/modules/polaris/polaris/requirements.txt
requests
sshtunnel
paramiko
pyyaml
uvicorn
aiohttp
psutil
rich
httpx
docker
loguru
GitPython
questionary
python-dotenv
click==8.1.3
tabulate==0.8.10
click-spinner